refactor(ooda_loop): extract run_ooda_cycle into cycle.rs (#1266)
mod.rs 442 -> 65 LOC; new cycle.rs ~398 LOC.
